main_project/requirements.txt
asgiref==3.5.2
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.2
drf-spectacular==0.26.5
django-debug-toolbar==4.2.0
sentry-sdk==1.35.0
urllib3==2.1.0